Bump dependency to libgtk-java 2.8
Unfortunately there are two non-backwards-compatible change as of GTK
2.8: The first is the introduction of glib-java and the fact that its
jars must be added to the CLASSPATH. The second is the removal of the 
Handle constructor use to get around the denaturation problem.

This patch adds the glib-java dependency to configure across Gentoo,
Debian and Fedora (although at time of writing the transition to
glib-java was not yet released) and switches to casting Widget rather
than using Handle constructors.