Rename action API methods in Command
Improve the method naming in the Command hierarchy. If you write a
Command subclass, then you now have to override (implement) persist()
which is a better name than execute(). And, if you're using a Command
class, then externally you call .execute() as the publicly visible
"get to it" method, which lines up better with Command pattern usage.