Move UnitOfWork from Command constructor to execute() method
In order to make subclassing Command cleaner, pass the UnitOfWork in
as a parameter of the execute() method instead of making it a soft
requirement of each Command constructor. This works a bit better as
the UnitOfWork is only actually needed by a Command the moment it
persists something; it doesn't have interest in objects, and it
doesn't try to save things in bits and pieces along the way. Command
subclasses must now implement persist(UnitOfWork) instead of
persist().