Reduce EditorWindow subclass of GladeWindow to editor only
EditorWindow used to describe windows that were both viewers and
editors and which would toggle between them, manipulating sensitivity.
This complexity is all removed. Now EditorWindow just describes
editors. Methods to be overridden  are now limited to cancel() and
ok() only. It is expected that in these methods UnitOfWork.cancel() or
UnitOfWork.commit() will be used.

Renamed to class and glade file to ReimbursableExpensesEditorWindow,
fixing a spelling mistake in the name and better conforming to
hierarchical class naming. Glade file's top level window now
ReimbursableExpensesEditor_window