Cache Books lookup; permit Command chaining
InitBooksCommand couldn't be used in a chain with AddAccountCommand
because DataStore.getBooks() went to the database, causing a
NullPointerException because no Books was yet commit()ed. So add a
setBooks() to DataStore, and in getBooks() cache the
lookup. Add a variety of exceptions to enforce proper usage.

New unit test class, CommandsVsUnitsOfWorkTest, with a three tests
observing this behaviour and ensuring cached Books object. Minor
changes to BasicCommandTest to match.

OprDynBooksSetup program reverted to not having to use double
UnitOfWork which was the point of the exercise.