Implement parent Ledger check in PostTransactionCommand
Major patch. In the process of adding a check in
PostTransactionCommand's action() method, realized that a considerable
amount of parentXXX machinery in Entry was missing or borked.

Now, to construct an Entry you need to specify an Amount and a parent
Ledger (note not Account) to which the Entry will belong *when
committed*. To facilitate existing unit tests, a null argument for the
Ledger is acceptable, though obviously PostTransactionCommand will
throw CommandNotReadyException if that is the case by execute() time.

Assumption change: the Entry constructor no longer sets the Datestamp
to be today as a default. Fixed a few unit tests to keep behaving as a
result, but in practice any Entry used in the system will become part
of a Transaction and will have it's Date set then.