PostTranasctionCommand adds each Entry object to specified Ledger
Entry objects are the bridge between Transaction and Account's Ledger.
An Entry must have it's parentLedger set before passing it as a part
of a Transaction to PostTransactionCommand; now Ledger's addEntry()
method [which actually adds the Entry's value to the Ledger's balance]
is called in PostTransactionCommand's action() before the Transaction
is committed to the underlying DataStore.