Expose query interface on DataStore
Prepatory to adding mock transactions into the demo database, we need
to expose a way to fetch objects. Db4o obviously provides that, but
this needs ot be wrapped by DataStore. So add a query() method to do
this, expecting more methods to come.

Add unit test. While at it, the unit tests which evaluate persistence
had been replying on a test class which was burried in one of the
TestCase classes. DummyInts is now in it's own file, making it easier
for other unit tests to use.

The form of Transaction which takes an array of Ledgers has Datestamp
added with the aim that these can be constructed in a mock up
situation and successfully passed to the Post command.