Add NotActivatedException; modify Datestamp to use
There are cases where db4o's automated activation will fail to
penetrate sufficiently with the result of default or null member
fields. NotActivatedException is to be thrown by code which recognizes
this condition.

Datestamp has a default value of UNSET (-1) which differs from the
default unactivated long value of 0, allowing us to detect an
unactivated state.

Also remove the _ in Datestamp's field.