Add AccountComparator and EntryComparator, use in toOutput()
In order to display accounts in a reliable order suited to accounting
purposes, and to order groups of entries when listing or dumping them,
create AccountComparator and EntryComparator.

The usual "avoid the urge to sort" and use a TreeSet to pre-sort your
data does not apply here - we have an out-of-band index care of our
persistence mechanism, so we leave the LinkedHashSets alone, instead
constructing a TreeSet(Comparator) when necessary. This is done in the
toOutput() routines presently in Books and (through Account to)
Ledger.

Moderately careful crafting was necessary in EntryComparator to ensure
that two objects that are outwardly identical (yet distinct) do not
get reduced to unity by the Set operators [which, when supplying a
Comparator, uses it's result as equals()]

Some fine tuning of the formatting in Ledger and Entry.toOutput().