Redo db4o config to cascade; remove Db4oSet; add activation unit test
Major patch. Finally get db4o querying working and to make
"transparent" the activation of objects from the database and updates
to make objects persistent:

Complete redo of the db4o configuration in DataStore. Two groups of
classes are tuned: classes which must cause a cascade of {activation,
updating}, and classes which are final leafs in the object graph such
as Datestamp and Amount.

Removed Db4oSet. This has been in the works for a while; all attempts
to use their magical database backed Maps caused race conditions.
Fixed by tuning the cascade settings of LinkedHashSet (the collection
used in all the domain classes), upping the default update depth
slightly, and deprecating attempts to replace the internal Sets
in-flight.

Considerable reduction in calls to UnitOfWork.registerDirty() down to
one per Command only, letting the default update depth along with
cascading update in the LinkedHashSets work magic instead. This method
will probably refactor to persist() or store(), leaving
registerDirty() for signalling.

Wrote an extensive unit test, Db4oActivationTest, which mimics the
combined SetupDemo, TransactionsDemo and OutputDump activity in order
to detect configurations which fail to do sufficient activation. This
unit test is also runnable from the command line with an argument to
specify the number of internal accounts - allowing external stress
testing.