Add toOutput() to Transaction; major change to EntryComparator
Entry's toOutput() already had some notion that it could be listed
either within a Ledger or within a Transaction, but the later side was
unimplemented. Print account and ledger in that case.

Major change was necessary to EntryComparator to deal with the
multiple contexts possible. Now you construct it by passing an object
(this will do) representing the context within which you're comparing,
either Ledger or Transaction. The sort critera are different depending
on which is used.

Transaction needed a getClassString just as Account and Ledger
hierarchies had. Implemented there and in GenericTransaction.

With these modifications, a toOutput() method was able to be added to
Transaction. While the encapsulation is nice, I think it's about time
this stuff moved out to a new set of classes. Bit too much
presentation to be stuck in the domain layer.