Create unit tests for Comparators; add hashCode differentiation
Added ComparatorsTest to accounts.services. Although normal usage had
validated the three comparators, the fact that TreeSet can swallow
(reduce) the number of evident objects if the comparator thinks they
are the same (returns zero) means that we needed a unit test to be
sure.

In practice, the existing logic was sufficient, but test cases created
a situation where too much was the same, and so following the example
in TransactionComparator, added Object.hashCode() as the final
differentiator in AccountComparator and EntryComparator.