Remove toOutput from domain classes; create TextOutput hierarchy
As a basic debugging measure there was a recursive toOutput() chain in
the domain classes. Needless to say, as presentation (even if just a
simple textual dump form) it didn't belong in accounts.domain.

Create TextOutput and subclasses TransactionTextOutput,
AccountTextOutput and EntryTextOutput. The Account one handles the
Ledger level. The Entry one switches its behaviour depending on the
context (listing Entries from within a Ledger or a Transaction).

TextOutput carries some static routines to handle padding and aligning
output, as well as Alignment constants LEFT and RIGHT. There's also
chomp() to truncate and add "..." as elipses.

TextOutput calls new class generic.util.Environment to ask for the
width of the terminal screen. Ideally this would actually get the
environment variable, but modern JDKs make this difficult and in any
case it doesn't seem to be exported to the environment by bash. So
wrap getting a property by the name of COLUMN. Add a resize exec to
the Makefile so provide this value in `make dump`. Failing all this,
TextOutput has a default width of 80.

This COLUMN value is then used to set the widths for the various
pieces of output text generated by the various subclasses of
TextOutput's toOutput() methods. The wider the terminal, the more room
for various displays. Note that Account's code is, at the moment, not
bothered with.

Finally, the sample dump program presently named OprDynOutputDump is
altered to now run a AccountTextOutput and TransactionTextOutput on
the demo data.