Update ObjectiveAccounts main file to include copyright; add JavaDoc
Add a public static field to the "global" ObjectiveAccounts class to
document the copyright for this code (such that it ends up in the
compiled bytecode and is quickly usable from about dialogs and
whatnot) and a reference to the terms of use in the top level licence
file.

Add JavaDoc for the public static variables in this class.

Fix LICENCE file to include 2006 in copyright statement.