Add Identifier, IdentifierGroup, and test suite
There is a need for the notion of a "constant" able to be persisted to
the data store - not immutable, but almost with primary key type
longevity. This is implemented in Identifier, added. Also added is
IdentifierGroup which holds a collection of such objects maintaining
their ordering.

Add basic unit test.

Leverage existing IdentifierAlreadyExistsException, but move it to
accouts.persistence as it is used by domain classes but not a domain
object itself.