Add customized JUnit textui VerboseTestRunner and VerboseResultPrinter
The normal JUnit textui TestRunner prints dots for test cases which
run. Desiring a bit more detail when `make test` is run to reflect the
effort put into unit tests, print the name of each test case on a line
and then ok, failed or error.

Add two new classes: VerboseTestRunner, which is more or less a
standalone simplified reimplementation of junit.textui.TestRunner, and
VerboseResultPrinter, likewise a simplified reimplementation of
ResultPriner which uses our TextOutput routines to keep things nicely
formatted.

The ui will halt on the first failure or error, but that is
controllable via VerboseResultPrinter.haltOnBug which is set from the
constructor.

Change AllTests to use this new code, but also the Makefile test
target to run it directly.