Add StoreObjectCommand
Sometimes you just need to quickly persist something; obviously that's
the whole point of db4o but we wrap this in the UnitOfWork and Command
framework to provide abstraction, isolation, and signalling.  So add
StoreObjectCommand in order to provide trivial persistance in the rare
cases when there is no domain validation to be done.