Add Finder base class
As long anticiapted, we introduce Finder as the base class of a
(otherwise flat) hierarchy of classes which contain specialized use
cases to retrieve specific objects from the database.

query() is the method that Finders will implement in common. It (and
presumably other Finder subclass methods) throws NotFoundException,
newly added for this case.