Infrastructure for storing Australian PAYG tax table data
Second part of the major patch series supporting Australian payroll
tax calculation.

In order to do the calculation the algorithm needs a set of
coefficients; the Identifier system now gives a way to index these,
but also needed was the means to retrieve the appropriate set of
coefficients given a specific target date. static { } data in the
Calculator class, though tempting (after all, it's just a bunch of
doubles) didn't really meet the lookup needs and more importantly the
ability to store multiple variants of data aggregating over multiple
years. And it looked ugly.

So create AustralianPayrollTaxTable to store these coefficients and
associated metadata in the datastore, and create a Finder subclass
AustralianPayrollTaxTableFinder to fetch up the data. The query()
method is implemented, but the heart of the matter is getCoefficients
which uses the query() result to then find an appropriate set of data
for the specified date.

Of course, the actual code specifying these coefficients had to go
somewhere, and it took a long time to find a place for them that would
hold up not just for initial creation, but also as they evolution and
aggregate over time.

To achieve this, we introduce a new kind of Command, one which
contains the data to be stored, rather than acting an agent for data
passed in to it. What makes the idea even more novel is that this
Command checks for previous instances of the data it stores, updating
those references if found.

The new Command is StoreAustralianPayrollTaxTablesCommand, which
presently contains two of the seven or so tax tables for 2005, proving
the concept.

Finally, update AustralianInitBooksCommand to call this Command after
setting up the identifiers.