Add AustralianPayrollTaxCalculator and unit tests
Third and final patch in the series to add a mechanism to calculate
the PAYG owing on salary payments made to workers in Australia.

One of the problems for a small business is that a certain amount
might be paid to a Director or Employee, but since payroll taxes need
to be withheld (and remitted to the tax office) a notional salary
needs to be worked out that was then "deducted" from to arrive at the
paycheck actually given.

This is the use case covered by calculateGivenPayable() in
AustralianPayrollTaxCalculator. This method, and the more conventional
calculateGivenSalary() are specified in PayrollTaxCalculator, which
inherits the notion of a specific date at which a payment happens and
at which taxes need to be worked out from the new base class
TaxCalculator.

AustralianPayrollTaxCalculator uses the coefficient data fetched from
a datastore via AustralianPayrollTaxTableFinder, indexed by the
AustralianPayrollTaxIdentifier instances established by
AustralianInitBooksCommand.

Extensive unit tests exercising the TaxCalculator hierarchy, the
finders, and the initialization and storage of Identifier+Datestamp
indexed coefficient data by StoreAustralianPayrollTaxTableCommand are
contained in the new test case file AustralianPayrollTaxCalculator