Change from LD_LIBRARY_PATH to -Djava.library.path in Makefile
Small change, but makes the execution environment a bit tighter; where
the VM looks for native (ie JNI) libraries can be specified by a
system property. This means we don't need to play games with
LD_LIBRARY_PATH which can cause broader problems when trying to bring
up the VM in the first place.