Add AustralianPayrollEditorWindow, PayrollTransaction, and TwoColumnTable
Major patch. Introduce PayrollTransaction to convey a salary payment
to an Employee into the accounts.

Create AustralianPayrollEditorWindow, a programmatic version
EditorWindow. User can enter either a salary amount or a paycheck
recieved and the other field, and withholding, will calculate.
Changing the selected AustralianPayrollTaxIdentifier will cause a
recalculation based on whichever field was most recently changed.

Both of these classes need work, in particular around the formation of
the PayrollTransaction (the Ledger lookups are hardcoded) and Employee
(which is mocked up and needs a Picker), but they are functional
within the limits that DemoWindowRunner exhibits. Also salary
calculation is just for one week so needs to be improved to calculate
time period elapsed and appropriate withholding.

In order to layout the fields of the EditorWindow in a tabular
fashion, create helper class TwoColumnTable, a subclass of GTK's Table
class which provides a simple mechanism for adding widgets to left or
right sides of a two column Table.

Generic UI element IdentifierSelector created based on design of
CurrencySelector Used in AustralianPayrollEditorWindow to allow
picking amongst PAYG tax scales.