Add AmountEntry and AmountDisplay
Create two new UI elements, AmountEntry and AmountDisplay. The first
allows a user to enter an Amount, the second displays (wrapping a
Label) an Amount in simple unformatted form.

Introduce interface accounts.ui.ChangeListener whose userChangedData()
is explicitly called once as a result of the event storms which happen
when an Entry is updated. This allows a Window with a (say)
AmountEntry to get a callback when a change occurs as a result of user
activity directly on that widget.