Create generic.ui.Master as a central re-entry point
Remove static reference to accounts.client.ObjectiveAccounts from
AbstractWindow now in generic.ui. That import was there just to get to
ObjectiveAccounts.abort(), so refactor this usage to new class Master.

Add the notion of a Hooks class of callbacks which can be added by registering
them with Master which will in turn be called by Master.abort() or
Master.shutdown().