Add WorkerPicker widget
Create a proper WorkerPicker widget, delegating to a ComboBoxEntry.
This widget is suitable only for small lists, not having the power of
AccountPicker. WorkerPicker enforces that the value of the entry field
must be one of the Objects stored within its internal ListStore.

Replace mockup use of ComboBoxEntry in AustralianPayrollEditorWindow
with WorkerPicker.

Update DemoMockTransactions to create a pair of Employees for testing
purposes. For now use StoreObjectCommand, but this will need upgrading
to some sort of AddEmployeeCommand. Add a mock constructor to
Employee.