Remove getLedger() from DataStore
Now that we have SpecificLedgerFinder, remove getLedger() from
DataStore. Adjust the things that were using it (DemoMockTransactions
and a few unit tests) to use the finder code instead.

As the original getLedger() code was an excellent example of a db4o
SODA query, preserve it (within a JavaDoc <pre> block) in
SpecificLedgerFinder.