Change Command to use DataClient; deprecate UnitOfWork
Major patch. The UnitOfWork notion was a good one to provide an
application level of continuity through transactions and an update
signalling framework. With our switch to a client/server mode of
accessing db4o, however, we can no longer rely on referential identift
between clients. So switch Command.execute() to operate directly on a
DataClient, and deprecate UnitOfWork entirely. This works out
reasonably nicely as UnitOfWork was mostly just in the way as an extra
wrapper around .save() anyway. We will need to reimplement the
signalling mechanism in due course, this time between DataClients.

Update InitBooksCommand, AddAccountCommand, and StoreObjectCommand to
use action(DataClient) as a preliminary check of concept. Rearrange
unit tests to evaluate Engine first, then basic Commands around Books. 

Create BlankDatafileTestCase to automatically create and open a blank
database to streamline unit test code which can inherit this behaviour
and only need to specify the filename to work with.