Refactor Commands and Finders to use DataClient instead of UnitOfWork
Major patch: continuing the series of patches migrating away from
UnitOfWork to db4o's client/server structure upgrade remainder of
Commands to use DataClient.

Upgrade remainder of unit tests to use new system, and migrate many of
them to use BlankDatafileTestCase to provide basic harness setup.

Upgrade Finder and its subclasses to use new API. Necessitated
exposing a call to query(DataClient) as a mandatory step in using
Finders. Alas.

Fix a bug in AddEntityCommand which allowed a null Ledger name to be
stored intead of using the Entity's name, which was the whole point of
that Command.