Create AddWorkerCommand
Create a specific command to add a Worker to the system, doing
appropriate validation and ensuring that a related Ledger is created
for him/her/it in the ReimbursableExpensesPayableAccount.

Add a unit test for Worker Commands.

Move the Employee and Entity commands from DemoMockTransactions to
DemoBooksSetup, and switch to using the new AddWorkerCommand.