Refactor PostTransactionCommand and add UpdateTransactionCommand
Create TransactionCommand as supertype of PostTransactionCommand and
new UpdateTransactionCommand which takes an existing Transaction,
peeks it to carefully find out what Entries are no longer present in
the live version but which are persisted, then removes all the Entries
(to update balances), [re]adds current Entries, saves new Transaction,
and delete's removed Entries.

Minor cleanups (guards, mostly) in Ledger and subclasses.

Modify TransactionCommandsTest with a unittest to verify correct
behaviour of UpdateTransactionCommand.