Update DemoMockTransactions to include Reimbursables and Payroll
To facilitate the testing of ability of various EditorWindows to
revise transactions, and to have a richer list of mocked up
Transactions, add ReimbursableExpensesTransactions and
PayrollTransactions.

Add artificial mockup constructor for PayrollTransaction, specifying
that the Entry[] provided must be in a specific order of salary,
withhold and paycheck.

Search up a bunch of Currency objects from the database (changing its
no arg constructor to public) and use those to make sure we have a
wide range of ForeignAmounts in the mocked up data.