Create TransactionListView and demo window to display it
At long last, create a widget to display a list of Transactions.
TransactionListView is (for the moment anyway) a TreeView subclass,
which presents a view of a Set of Transactions, one row per.

Bulk of work to define presentation is in a populate() method staffed
with the ListStore.setValue() calls. It was necessary to go to
considerable pain with the Entries to square them to be equal width so
that lines in CellRendererTexts after \ns would appear to be right
justified.

Create a temporary Window called TransactionListViewHolder so
DemoWindowRunner has something to pop up.

Still trouble with the Debit/Credit colours - when a row is selected
foreground colour reverse videos, but not text coloured with explicit
pango markedup.