HIG improvement for ForeignAmountEntryBox; gray out widgets as group
Add 'x' and '=' to ForeignAmountEntryBox as a usability (or at least
an understandability) improvement. Thanks to Maciej Katafiasz
<email@mathrick.org> hackers for the suggestion.

For whatever reason, although Eclipse is happy to parse and compile
unicode characters, when Jikes goes after them it compiles but the
result is a leading Ao then either the character or the
unicode-unrecongnized block symbol. So, to keep things peaceful, I've
used the \uXXXX notation in the source code itself.