Slight colour improvement for TransactionListView selected row
Pick slightly darker colours for Credit and Debit so that they appear
a bit clearer when in selected row in a TreeView.

Improve TransactionListView.java to ensure & characters are escaped if
present in Account title or Ledger name. Use charcoal and darkgray
respectively as adjustable constants for colouring separator and
currency codes.