Have DataServer refresh objects fetched by DataClient on release
As DataClient queryByExample() and nativeQuery() are called, quietly
build up a list of the objects that have been so fetched.

Then, as a measure of protection against stale data, when a DataClient
is returned to the connection pool, the DataServer will iterate over
the DataClient's list of dirty objects and call ExtObjectContainer's
refresh() on these objects to return them to their READ COMMITTED
state.

Update the unit test to verify that this behaviour is actually being
experienced, removing a test for wacko behaviour that not doing this
permitted.