Customized colours for when rows are selected in TransactionListView
Since we use blue and green for Debits and Credits, respectively, the
blue gradient used in the selected row bar in GTK TreeView widgets
blots out these colours. So split colours in Debit and Credit to be
_NORMAL and _ACTIVE, upgrade Account and Ledger's getColor() to use
them (by feeding in a boolean).

Refactor TransactionListView to [re]populate a row according to
whether it is active or not. Add a TreeSelectionListener to call
setAsActive() with a TreePath indicating which row is selected, then
have populate() act accordingly.