Add Class parameter to Engine and DataServer methods
db4o has to have its configuration set before you open up an
ObjectServer. We do this in a static block in Books, but there were
bugs because this wasn't getting loaded before DataServer.

So, load it. Add a parameter to DataServer's constructor (and the same
parameter to the two Engine class methods which call it) to pass a
Class object representing the concrete subclass of Root which contains
the initialization block. [in our case, Books.class]

Use null in the more basic unit tests which don't have to worry about
any of this.