Enable activation of an Editor from TransactionListView
Changes across the board to allow an activate event in
TransactionListView to result in an EditorWindow being raised.

In particular, UpdateTransactionCommand didn't seem to work after all.
Redo the logic there by improving the query used to retrieve the
committed Entries. This necessitated adding an equals method to
Transaction which doesn't evaluate referential equality.

Adjust AustralianPayrollEditorWindow to have a no-arg constructor (for
new) and a constructor with a long id (for editing) as a cross
container way of summoning the correct Transaction up for editing.
Adjust the way references are assigned from the passed in Transaction
so that things are set as needed before widgets render.

As a minor tuning in EditorWindow, do the release of the DataClient in
a Thread so that the UI can get on with hiding the Window to improve
apparent responsiveness.

Further improve DemoMockTransactions to use a real Identifier when
adding a PayrollTransaction.