Enable editing of existing ReimbursableExpensesTransactions
Fix up ReimbursableExpensesEditorWindow to be able to edit an existing
transaction. Implement the same API as with
AustralianPayrollEditorWindow whereby a no arg constructor means
create a new Transaction, whereas the <init>(long) constructor means
edit the Transaction with that database id.

AccountPicker needed some refactoring so that setAccount() worked
properly; add setLedger(). Likewise fix the order that subwidgets are
set in ForeignAmountEntryBox to ensure references are maintained.
Clean up some fluff in CurrencySelector.

Move the present() call out of constructors of
ReimbursableExpensesEditorWindow and AustralianPayrollEditorWindow and
have it done from whatever is instantiating the window, ie
ObjectiveUserInterface's launch() method.