Update TransactionTreeView when EditorWindows close.
Tigthen up the Master.ui.launch() API to take a DomainObject as its
argument reducing the parameter clutter. Applications (ie our
ObjectiveUserInterface) now implement launchEditor(long, Object) which
is to return the EditorWindow instantiated.

Push the EditorWindow-to-ID mappings up to UserInterface.

Reuse UpdateListener (an artifact from the old UnitOfWork) as the
observer interface called when an EditorWindow closes to signal
Widgets capable of rediplsaying the data they're showing to do so. New
interface method redisplayObject(long). Widgets with interest have to
register it by calling Master.ui.registerListener(UpdateListener).

Move UpdateListener and EditorWindow to generic.ui.

Finally, when an EditorWindow changes the Transaction it is editing,
the result is propegated back to TransactionTreeView. Refactor its
populate() code so that the "is this row active" parameter is now
stored in a column of the ListStore so that redisplayObject() can just
find the right TreeIter and tell populate() to redraw it.