Extend Identifier with an index field; use to fix AustralianPayroll tests
Add an optional index field to Identifier, propegating it down to to
AustralianPayrollTaxIdentifier's constructor. Use it so that tests and
demo code can reload the constants there.

Add a DomainObject field to Selector and a constructor which allows
you to pass a DomainObject as a query variable into a native query.