Use a worker Thread in ReimbursableExpensesEditorWindow.ok()
Move the Transaction update and save+commit code in ok() into a worker
thread and fire it off so that UI doesn't hang while this is going
on. This could probably be done a bit more cleanly, especially considering
it's probably something most EditorWindows are going to need to do.