Bugfix: Ensure selected row's remains highlighted on resort
Bug: when TransactionListView has a selected row, and then the user
tells it to sort on a different column, the wrong row is highlighted.

Cause: TreePaths and TreeIters aren't stable if the TreeModel they are
over changes - which happens when you tell a TreeView to sort on a
different column. 

Fix: Change previous field [as used by setAsActive()] from a TreePath
to a TreeRowReference.