Display start and end date in AustrlianPayrollEditorWindow
Add start and end Datestamps to PayrollTransactioni's mockup
constructor, and update DemoMockTransactions with some sample data.

Fix a bug in Datestamp where it was creating new objects rather than
updating the value of the existing one.

Display the start and end Datestamp in AustralianPayrollEditorWindow
separately from the Transaction date. This needs to migrate to a
range picking widget.