Switch to using DataColumnLong in TransactionListView
For three of the sorting DataColumns (Date, Debit, and Credit) the
lack of a DataColumnLong meant we had to jump through hoops to squeeze
the natural long values underlying these data into DataColumnInt
columns.

With the patch (forthcoming in libgtk-java 2.8.5) adding
DataColumnLong to the TreeModel type family, we can store the long
values (Datestamp's timestamp, and the long cents under Amounts)
directly in the sort columns.