Bugfix: ModalDialog needs to be told who the parent Window is.
MessageDialogs (as wrapped by ModalDialog) were just grabbing the
first Window  in the present Window array, with the result that there
was bad flicker as one  window was presented, then focus was returned
to the one (care of our explicit  present()).

Adjust MessageDialog's constructor to take an [org.gnu.gtk] Window
(generally available in any of our code as [AbstractWindow] window).