First take at an editor for GenericTransactions
Add GenericTransactionEditorWindow. Unlike other use cases, here the
user has to worry about allocating Amounts as Debits or Credits. So a
row (HBox) is presented for each Entry and buttons are provided to
allow one to convert a Debit to a Credit.

Adjust ObjectiveUserInterface's launch() to instantiate one of these
windows for GenericTransactions.