Improve GenericTransactionEditorWindow's flip flop behaviour
The Debit/Credit buttons were cumbersome. Change to a single button
spanning both Debits and Credits columns, so that there is a larger
target to whack if the user wants to flip from one to the other.

Also fixes problem with button staying normal after focus was moved
out from underneath it.