Add updateEntry() to Ledger
Add an updateEntry() method to Ledger, complementing addEntry() and
removeEntry() so that surrounding code can tell the Ledger than an
Entry has had its Amount changed and thus the Ledger needs to
recalculate its balance.

At the moment this is somewhat of a hack. The method is essentially a
no-op; all it does is null out the cached balance Amount. A comment
indicating that a better algorithm needs to be used is included.