Minor bugfix to DatePicker
Resolve a minor glitch whereby GTK was failing an assert test. Cause
was setting the day before setting the month of the GtkCalendar, with
the result that if setting day 31 and there wasn't a 31st day in the
current month, it would complain.

While we're at it, remove the leading _ from class fields in
DatePicker and its DatePickerPopup, and a javadoc improvement to
Datestamp.