Add summation AmountDisplays to display net Debit and Credit balances.
In order to provide visual feedback about whether or not a
GenericTransaction is in balance, add AmountDisplay widgets to display
the total of the Debits and the total of the Credits present in the
Transaction being edited by a GenericTransactionEditorWindow.

Further complicating an already convoluted Window, this is done by
using the existing code in Ledger to sum Entries taking their side
into account via two mock Ledgers, one for each balance being
computed.