Redo UpdateTransactionCommand
Continued diffiuclty with db4o read concurrency led to reimplementing
UpdateTransactionCommand in an attept to have a cleaner algorithm,

I suspect that in the end we ended up with much the same as we started
with, but it does read more smoothly. Along the way, replace the
complex native query aimed at retrieving the ledger containing a
specific Ledger, and instead just do a graph walk along the Ledgers,
and use contains().

Change DomainObject's congruent() to not throw exceptions, but rather
to just return false when things aren't quite right. This probably
didn't impact UpdateTransactionCommand's native queries after all, but
nevertheless it seems a good idea to not throw RuntimeExceptions in
code that is buried in match() implementations.

Add a third test block to testUpdateTransaction() in
TransactionCommandsTest to go over the logic of removing entries and
re-adding them - which is what GenericTransactionEditorWindow does -
and make sure that the underlying infrastructure doesn't choke when
this happens.