Redo AccountPicker, creating AccountLedgerDisplay
Reimplement the AccountPicker widget, changing it from an Entry with a
tiny helper Button to a single wide Button with dynamic text
indicating the selected Account and Ledger.

The widget used to render the text of the Button is new class
AccountLedgerDisplay. This is a Label whose text is the Account title
and Ledger name properly marked up with Debit/Cridit-ness colour.

AccountLedgerDisplay has an ExposeListener which calls a fairly
complex resize() method. This code is carefully crafted to evaluate
the amount of space allocated to the enclosing HBox, the amount of
space the Label wants (requests), and then re-rendering the label with
a shorter width. This iterates until the Label fits in the Box, or
until the (very small) minimum width is reached.

The code to markup the Account title and Ledger name with a given
restricted width (in characters) is in a static class method called
markupTitleName(). This code is originally that first used in
TransactionListView.populate(); there is still the variation of
active/inactive to consider so a proper refactoring of that code out
of TransactionListView will wait.