Add comment indicating possibility of using server's ObjectContainer
It turns out that you can get at the ObjectContainer being used by the
server side of a client/server db4o setup. This offers the possibility
bypassing the network (or at least, inter-thread) overhead.

This is very tempting to use for Engine.primaryClient(). In the unit
tests, doing it this way shaved 20 out of 67 seconds required run
time. Encouraging! And, it appeared to work fine for the application
as shown by DemoWindowRunner.

There's a huge risk though - this is messing with the server
internals. In particular, it's difficult to forecast what the impact
of the UI calling reload() on a particular object graph might be. So
until we better understand the code paths involved, we'll leave it out
for now.