Generate javadoc with -linkoffline instead of -link
Setup rules to fetch the package-list files for each of the libraries
we're linking against, caching them in tmp/javadoc so that doc
generation doesn't rely on a network connection.

javadoc itself did do the fetching, but it keeps re-fetching, which is
silly. Obviously we'll usually have network access for this but it
means that we can rebuild javadoc if briefly disconnected so long as
a successful fetch has happened already.