Resolve bug with half calendar vs half year
To great annoyance, the previously calculated $572 withholding for
$6000 earnings in 6 months was not correct since we used 26 weeks not
the 26.2 weeks between 1 Jul 05 and 31 Dec 05.

After adding a unit test and working through it all, discovered that
correct withhold for 26.2 weeks is $550. Salary of $6572 does work to $6000
paycheck if weeks is 26[.0]. So fix up unit tests.

Along the way, improve algorithm in AustralianPayrollTaxCalculator
with respect to rounding and no longer needing BigDecimal.

Update DemoMockTransactions with correct initialization data.