Fix DatePicker to work in java-gnome 4.0
Complete upgrades to DatePicker, along the way exercising the
keystroke handling in java-gnome 4.0

Start cleaning up AbstractWindow only to realize that we really don't
need a bloody Glade file just to manage a single Calendar Widget.
Jeesh. I can't remember whether the AbstractWindow is used as a global
construct in .ui. management, but given that it was mostly about the
nastiness of doing LifeCycleManager correctly, I think we'll be a lot
better off ditching it entirely.

Wrote a simple wrapper to fire up Windows that don't need the database
behind them. That won't last very long.