Fix internal databaseId in DomainObject.
Change from internal callbacks in DomainObject to the new external
callback registry, as invoked by DataClient.

This is by no means complete; beyond being disatisfied with the
workaround I am not confident in this overall design. It feels shaky,
and the db4o external callback API is attrocious.

DomainObject's setID() should not be public!

Can we do away with DomainObject entirely now? Yes, if we move the id
logic to our Selector wrapper around Predicate.